python-examples/24-session-sandbox/pyproject.toml
[project]
name = "polos-session-sandbox"
version = "0.1.0"
description = "Session-scoped sandbox -- reuse a Docker container across multiple agent runs"
requires-python = ">=3.10"
dependencies = [
    "polos-sdk[anthropic]",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
]
